# [PyTorch][Vulkan] Add a matrix multiplication performance test binary and fix GPU latency measurement